move keysight file to prevent duplication in docs build
