+++ qcodes/instrument_drivers/Keysight/keysightb1500/KeysightB1520A.py
@@ -10,7 +10,7 @@ from .message_builder import MessageBuilder
 from . import constants
 from .constants import ModuleKind, ChNr
 if TYPE_CHECKING:
-    from .KeysightB1500 import KeysightB1500
+    from .KeysightB1500_base import KeysightB1500
 
 
 _pattern = re.compile(
