+++ qcodes/instrument_drivers/Keysight/keysightb1500/__init__.py
@@ -4,7 +4,7 @@ Semiconductor Parameter Analyzer.
 
 The structure of the module is as follows:
 
-  - The driver of the Parameter Analyzer is located in ``KeysightB1500.py``
+  - The driver of the Parameter Analyzer is located in ``KeysightB1500_base.py``
     file.
   - ``KeysightB1500_module.py`` file implements a class that is common for
     all the modules that the Parameter Analyzer may support.
@@ -15,7 +15,7 @@ The structure of the module is as follows:
 
 """
 from .message_builder import MessageBuilder
-from .KeysightB1500 import KeysightB1500
+from .KeysightB1500_base import KeysightB1500
 from . import constants
 
 __all__ = ['KeysightB1500', 'MessageBuilder', 'constants']
